Added SignalProducer.replayLazily for multicasting
Motivation
====================

I expect this to be a controversial one, so let me provide some context: multicasting has always been an important feature.
Notably, *Rx* contains multiple operators ([replay and cache](http://reactivex.io/documentation/operators/replay.html), [multicast](http://reactivex.io/documentation/operators/publish.html), etc) for this. However, we've always been very cautious about including this in *ReactiveCocoa* **because it's very easy to misuse**. Specifically, the most common pattern that I see is this:
```java
final BehaviorSubject<Integer> subject = BehaviorSubject.create();
final Observable<Integer> observable = subject.getObservable();

// ... somewhere else in the code:

// This is safe because it's backed by a `BehaviorSubject`.
final int lastNumber = observable.toBlocking().first();
```

I've seen this **so many times**. It seems fine at first, but it's very important to realize that the comment "this is safe" is actually a lie: we have no compile time guarantees that it will not deadlock, or block the current thread for several seconds.
This is precisely why `PropertyType` was born: not only does it make this pattern a lot easier, but it provides compile-time guarantees and better semantics to implement this sort of thing.

This is exactly why I and others have felt that A: we don't need multicasting, and B: if we did add it, it would be used instead of using `PropertyType`.
However, I've ran into several scenarios where *replaying* was necessary. To accomplish it, I've been using `MutableProperty` (I even [suggested this in StackOverflow](http://stackoverflow.com/questions/33946571/how-do-i-observe-a-signal-and-immediately-receive-a-next-event-if-it-has-alrea/34010425#34010425)), so I thought it was time to provide a first-party solution to this.

Details
====================

I've been putting a lot of thought into this, whether it really belongs in RAC's codebase, the semantics that it should have, etc. And this is the result!
First let me explain what the semantics are (all of this is tested in this commit):

- Invoking the operator has no side effects, i.e. it's *lazy*. This is important to maintain consistency with the rest of `SignalProducer` semantics: only `start` can have side effects.
- Upon the first call to `start`, the underlying `SignalProducer* is started.
- Values are buffered to all subsequent subscribers, from that point on.
- This is slightly weird because the `SignalProducer` no longer has *value semantics* (its buffered values are shared). But this is already how things like `SignalProducer.buffer` or `PropertyType.producer` work: they also have reference semantics, so I think it's fine. It just means that starting the producer *can* have side effects, but most importantly, different observers will see different versions of the producer depending on when they `start` it.
- When the producer *fails*, the `error` is automatically forwarded to any subsequent observer.
- When the producer *completes*, new observers will see the last `n` values emitted, and then complete.
- The subscription to the underlying `SignalProducer` will stay alive even if the last observer disposes of it. This is important, because there might be new observers later on.
- If there are no more subscriptions to the replayed `SignalProducer`, and it goes out of scope, the underlying producer is disposed.
- The result of this operator **is another `SignalProducer`**: this is important because, as explained above, each observer will see a different version of the truth.
- This operator **is only implemented for `SignalProducer`**: this is also very important (and something that Rx can't do). it wouldn't make sense to multicast a `Signal`!

So notice that this is exactly the same semantics as `SignalProducer.buffer`. It allows you to multicast elements from a single source.
Thanks to this, its implementation is actually 100% based on `SignalProducer.buffer`!

I tried explaining these points as concisely as possible as part of the *dosctring*, but let me know if it could be improved.

Use cases
====================

- As an optimization: one can use this operator to avoid doing the same work multiple times. I needed this myself to get these semantics [in my rendering framework](https://github.com/NachoSoto/AsyncImageView/blob/master/AsyncImageView/Renderers/MulticastedRenderer.swift), and using an operator would be a lot easier.
- In cases where the same producer is needed multiple times as part of a complicated operator chain. For example, being able to do things like `combineLatest(p.filter(f1), p.filter(f2))` without incurring in 2 subscriptions to the same `SignalProducer`.

Questions
====================

- I'd like @jspahrsummers to take a look at this if possible. He actually taught me a lot of what I explained here regarding the common cases that lead to misusing *multicasting*. **I think** I've managed to come up with an implementation that minimizes the room for errors, and the *docstring* explains when `PropertyType` would be preferred, but I'd like to hear his thoughts, and anybody else's, regarding this controversial operator.
- I'm also interested in hearing other people's ideas for cases where they wished *multicasting* were easier, and make sure that these are the right semantics we want.