Actually added several UIKit extensions to the tvOS target
This is a follow up to #2651. Did I mention that I hate `Xcode`? Even though I added the files to the `tvOS` target, and they're imported in `ReactiveCocoa.h`, these extensions aren't found in `RC 2`.
Why the hell is this not an error, for importing a non-modular header? I swear I understand Xcode less and less every day. Anyway, with this change, NOW their headers are `public`...