Allow RACCommand to specifiy covariant input type
Inputs to `RACCommand` to do not suffer from the same issues as `RACSignal` when attempting to adopt lightweight generics as discussed #2085. As such, adding them is as simple as declaring a `<__covariant InputType>` in the interface.

In projects that are still using RAC 2.X, this will be helpful to ensure that commands are executed with an input of the correct type. It also helps to catch programmer error when refactoring commands.

I was unable to find a way to expose this annotated type to the compiler when bridging a `RACCommand` to `Action` in `ObjectiveCBridging.swift`. Open to any suggestions on how this might be possible.