Add lightweight generics to signal operations where applicable
For many `RACSignal` operators, the `ValueType` of signals will not change due to operator being invoked. As such, in these cases, lightweight generics can be added to improve the type signature of the return value.

This additionally adds lightweight generics to RACEvent so that it can maintain the `ValueType` of a `RACSignal` bridged into it via the `materialize` operator.

Finally, in some places (such as class methods on RACSignal`, the return type has been updated to be `RACSignal<ValueType> *`. This is to allow the following style of declaration in Swift:

```
let signal = RACSignal<NSNumber>.createSignal { subscriber…
```

This follows the pattern of Apple’s generics, e.g. `+[NSArray (nullable NSMutableArray<ObjectType> *)arrayWithContentsOfFile:(NSString *)path]`