Redeclare operations built on RACStream primitives with more precise ValueType information
Since the operations built on RACStream primitives have an `instancetype` return value, the compiler assumes that they inherit the receiving signal’s `ValueType`. For operators like `map:`, this is almost always an incorrect assumption. As such, in cases where the `ValueType` of the result of the operation is not able to be inferred, the `ValueType` is erased in the result by returning `RACSignal *` with no `ValueType` information.

Additionally, in cases where `instancetype` is a valid return type and no more precise `ValueType` information could be added to the method signature, the operation is not redeclared, since the compiler will assume the correct return type.

Finally, in cases where more precise type information could be provided, the operators are redeclared as such.