Add warnings to unused signal operator results
One of the most common sources of confusion with developers that are new to the concepts of FRP is that subscriptions are required to drive values through signal operators. New developers frequently spend a long time trying to figure out why code such as:

```
[signal logAll];
```

is not performing any logging side-effects. To help with this confusion, I’ve added unused result warnings to side-effectless signal operators to help raise a warning to new-comers that operators are side-effectless.

This change will also help with catching typos that would otherwise compile without warning, such as forgetting to type `return` before a signal chain when you are attempting to write an early return in a method.