Add generic typed tuples
This change adds lightweight generic typed tuples as subclasses of `RACTuple`. This allows tuples to be annotated with the types that they contain, e.g. `RACTwoTuple<NSNumber *, NSString *>`. The supported counts are 2: `RACTwoTuple`, 3: `RACThreeTuple`, 4: `RACFourTuple`, and 5: `RACFiveTuple`.

To attempt to best satisfy the compiler, `RACTuplePack()` was updated to produce the correct `RACTuple` subclass depending on the number of arguments it is provided. In most cases, if you have a tuple with two elements, you will have a `RACTwoTuple` from now on (and so on). However, you can still produce a regular `RACTuple` with two, three, four, or five elements by using the explicit initialization methods (e.g. `RACTuple tupleWithObjectsFromArray:`).

In cases where you have a `RACTuple` with explicit and differing generic types, it is for the most part unnecessary to use `RACTupleUnpack()` now, since you can get type-safe access to elements by using `tuple.first`, `tuple.second`, etc. However, note that the subscripting methods (e.g. `tuple[0]`) are not possible to annotate with lightweight generic types.

Fixes #57