RACSignal+Operations: modify -sample: for more consistent behavior.
The `-sample:` operator completes once the receiver completes. This
behavior is not suited for this particular operator, which should more
likely complete once the both sampler and the sampled signals complete,
truly indicating it will not send any more values. Both signals still
pass errors through.

Moreover, the convention in RAC is that an operator that operates on
more than one signal completes when the all its inputs complete, and the
current behavior of `-sample:` breaks this convention.